Cambios de Exportacion, Armado de pales, Perfiles 04/12

+++ Sistema de Laboratorio/Sistema de Laboratorio/Maestros/Ingreso Frases H & P/IngresoFrasesH.vb
@@ -1,7 +1,8 @@
 ﻿Public Class IngresoFrasesH : Implements IFrasesHP
     ReadOnly FRASE As String
 
-    Sub New(ByVal TipoFrase As String)
+    Dim PermisoGrabar As Boolean
+    Sub New(ByVal TipoFrase As String, ByVal ID As String)
 
         ' Llamada necesaria para el diseñador.
         InitializeComponent()
@@ -23,6 +24,12 @@
 
         DGV_Frases.DataSource = tabla
 
+        SQLCnslt = "SELECT Escritura FROM PermisosPerfiles WHERE ID = '" & ID & "' AND Sistema = 'LABORATORIO' AND Perfil = '" & Operador.Perfil & "' AND Planta = '" & Operador.Base & "' ORDER BY ID"
+        Dim Row As DataRow = GetSingle(SQLCnslt, "SurfactanSa")
+        If Row IsNot Nothing Then
+            PermisoGrabar = Row.Item("Escritura")
+        End If
+
     End Sub
 
     Private Sub txtBuscador_KeyUp(sender As Object, e As KeyEventArgs) Handles txtBuscador.KeyUp
@@ -34,34 +41,36 @@
 
     Private Sub DGV_Frases_RowHeaderMouseDoubleClick(sender As Object, e As DataGridViewCellMouseEventArgs) Handles DGV_Frases.RowHeaderMouseDoubleClick
         Dim SQLCnslt As String
+        If PermisoGrabar Then
+            If (MsgBox("¿Desea borrar la frase " & DGV_Frases.CurrentRow.Cells("Codigo").Value & "?", vbYesNo) = vbYes) Then
 
-        If (MsgBox("¿Desea borrar la frase " & DGV_Frases.CurrentRow.Cells("Codigo").Value & "?", vbYesNo) = vbYes) Then
+                If (FRASE = "H") Then
+                    SQLCnslt = "DELETE  FROM FraseH WHERE Codigo = '" & DGV_Frases.CurrentRow.Cells("Codigo").Value & "'"
+                Else
+                    SQLCnslt = "DELETE  FROM Frasep WHERE Codigo = '" & DGV_Frases.CurrentRow.Cells("Codigo").Value & "' "
+                End If
 
-            If (FRASE = "H") Then
-                SQLCnslt = "DELETE  FROM FraseH WHERE Codigo = '" & DGV_Frases.CurrentRow.Cells("Codigo").Value & "'"
-            Else
-                SQLCnslt = "DELETE  FROM Frasep WHERE Codigo = '" & DGV_Frases.CurrentRow.Cells("Codigo").Value & "' "
+                ExecuteNonQueries("SurfactanSA", SQLCnslt)
+
+                If FRASE = "H" Then
+                    LblTitulo.Text = "Frases H"
+                    SQLCnslt = "SELECT Codigo, Descripcion = RTRIM(LTRIM(Descripcion)) + ' ' + RTRIM(LTRIM(DescripcionII)) + ' ' + RTRIM(LTRIM(DescripcionIII)), Observaciones = Observa FROM FraseH"
+
+
+                Else
+                    LblTitulo.Text = "Frases P"
+                    SQLCnslt = "SELECT Codigo, Descripcion = RTRIM(LTRIM(Descripcion)) + ' ' + RTRIM(LTRIM(DescripcionII)) + ' ' + RTRIM(LTRIM(DescripcionIII)), Observaciones = Observa FROM FraseP"
+
+                End If
+
+                Dim tabla As DataTable = GetAll(SQLCnslt, "SurfactanSA")
+
+                DGV_Frases.DataSource = tabla
+
+
             End If
-
-            ExecuteNonQueries("SurfactanSA", SQLCnslt)
-
-            If FRASE = "H" Then
-                LblTitulo.Text = "Frases H"
-                SQLCnslt = "SELECT Codigo, Descripcion = RTRIM(LTRIM(Descripcion)) + ' ' + RTRIM(LTRIM(DescripcionII)) + ' ' + RTRIM(LTRIM(DescripcionIII)), Observaciones = Observa FROM FraseH"
-
-
-            Else
-                LblTitulo.Text = "Frases P"
-                SQLCnslt = "SELECT Codigo, Descripcion = RTRIM(LTRIM(Descripcion)) + ' ' + RTRIM(LTRIM(DescripcionII)) + ' ' + RTRIM(LTRIM(DescripcionIII)), Observaciones = Observa FROM FraseP"
-
-            End If
-
-            Dim tabla As DataTable = GetAll(SQLCnslt, "SurfactanSA")
-
-            DGV_Frases.DataSource = tabla
-
-
         End If
+        
     End Sub
 
     Private Sub btnImprimir_Click(sender As Object, e As EventArgs) Handles btnImprimir.Click
@@ -109,7 +118,7 @@
                 For Each Row As DataGridViewRow In DGV_Frases.Rows
                     With Row
                         If Trim(UCase(.Cells("Codigo").Value)) = Trim(UCase(txtAccesoRapido.Text)) Then
-                            With New CargaDatosFrases(FRASE, .Cells("Codigo").Value, .Cells("Descripcion").Value, .Cells("Observaciones").Value)
+                            With New CargaDatosFrases(FRASE, .Cells("Codigo").Value, .Cells("Descripcion").Value, .Cells("Observaciones").Value, PermisoGrabar)
                                 .Show(Me)
                             End With
                         End If
@@ -120,10 +129,15 @@
     
     Private Sub DGV_Frases_CellMouseDoubleClick(sender As Object, e As DataGridViewCellMouseEventArgs) Handles DGV_Frases.CellMouseDoubleClick
         If e.ColumnIndex <> -1 Then
-            With New CargaDatosFrases(FRASE, DGV_Frases.CurrentRow.Cells("Codigo").Value, DGV_Frases.CurrentRow.Cells("Descripcion").Value, DGV_Frases.CurrentRow.Cells("Observaciones").Value)
+            With New CargaDatosFrases(FRASE, DGV_Frases.CurrentRow.Cells("Codigo").Value, DGV_Frases.CurrentRow.Cells("Descripcion").Value, DGV_Frases.CurrentRow.Cells("Observaciones").Value, PermisoGrabar)
                 .Show(Me)
             End With
         End If
     End Sub
 
+    Private Sub IngresoFrasesH_Load(sender As Object, e As EventArgs) Handles MyBase.Load
+        If PermisoGrabar = False Then
+            btnAgregarFrase.Enabled = False
+        End If
+    End Sub
 End Class
